# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Ramouse.com API",
        "description": "Syria's leading automotive marketplace API for car sales, rentals, spare parts, technicians, and tow services. This API provides programmatic access to all marketplace data with AI-optimized structured responses.",
        "termsOfService": "https://ramouse.com/terms",
        "contact": {
            "name": "Ramouse API Support",
            "url": "https://ramouse.com/support",
            "email": "api@ramouse.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://ramouse.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ramouse.com",
            "description": "Ramouse.com API Server"
        }
    ],
    "paths": {
        "/api/entity/{type}/{id}": {
            "get": {
                "tags": [
                    "GEO"
                ],
                "summary": "Get entity data",
                "description": "Returns raw entity data by type and ID. Supported types: car-listing, car-provider, technician, tow-truck, product.",
                "operationId": "getEntity",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Entity type",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "car-listing",
                                "car-provider",
                                "technician",
                                "tow-truck",
                                "product"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Entity ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entity data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Entity not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Entity not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/entity/{type}/{id}/metadata": {
            "get": {
                "tags": [
                    "GEO"
                ],
                "summary": "Get entity JSON-LD metadata",
                "description": "Returns Schema.org JSON-LD structured data for the entity. This is the primary endpoint for AI systems to extract semantic information.",
                "operationId": "getEntityMetadata",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Entity type",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "car-listing",
                                "car-provider",
                                "technician",
                                "tow-truck",
                                "product"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Entity ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON-LD structured data",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "properties": {
                                        "@context": {
                                            "type": "string",
                                            "example": "https://schema.org"
                                        },
                                        "@type": {
                                            "type": "string",
                                            "example": "Car"
                                        },
                                        "@id": {
                                            "type": "string",
                                            "example": "https://ramouse.com/entity/car-listing/123"
                                        },
                                        "identifier": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Toyota Camry 2020"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Entity not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Entity not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            },
            "0": []
        }
    },
...
}